add view, delete, delete all and reset agenda

+++ src/php/tool/todo/sql_todo.php
@@ -15,7 +15,10 @@
       return connect_sql($query);
     }
     function default_todo(){
-      $query="INSERT INTO `todo` VALUES (0,'Integer at odio porta, finibus mi vel, bibendum orci. Duis elit est, ornare in ornare eu, dictum et nulla. Pellentesque libero enim, consectetur sed lectus vitae, tristique interdum nisi. Suspendisse eget tempor erat, et cursus risus. Nam purus velit, facilisis a placerat tempus, consequat vitae mi. Nulla rhoncus, quam non imperdiet malesuada, erat enim sollicitudin nisi, in interdum orci mauris sit amet leo. Donec sagittis ligula ipsum, et pharetra augue porta ut. Donec vel posuere quam. Phasellus facilisis ligula a sem pretium, eu consequat nibh mattis. Fusce id turpis eget elit maximus laoreet eget nec turpis. Duis id sem lobortis, ornare turpis in, tempor augue. Donec at magna vel leo varius ullamcorper non ut elit.',0),(1,'Nulla facilisi. Maecenas tristique dolor in enim porttitor laoreet. Duis venenatis tincidunt purus, sed suscipit lorem rutrum ac. Proin mauris massa, blandit sit amet porttitor fermentum, pharetra sit amet diam. Aliquam id pulvinar orci, id condimentum lorem. Interdum et malesuada fames ac ante ipsum primis in faucibus. Sed elementum risus tellus, nec pharetra tortor feugiat id. Fusce pretium tincidunt efficitur. Praesent ut sodales neque. Duis imperdiet, risus vitae commodo fringilla, metus nibh egestas lectus, sed congue lorem enim nec ex. Duis nec porttitor diam. Donec aliquet metus eu ante cursus tempor. Nunc malesuada mi venenatis, tincidunt est suscipit, tempor nunc. Sed scelerisque erat id magna dictum, ut posuere mauris sodales. Suspendisse interdum posuere tellus, ut commodo metus sagittis ut.',0),(2,'Nulla facilisi. Maecenas tristique dolor in enim porttitor laoreet. Duis venenatis tincidunt purus, sed suscipit lorem rutrum ac. Proin mauris massa, blandit sit amet porttitor fermentum, pharetra sit amet diam. Aliquam id pulvinar orci, id condimentum lorem. Interdum et malesuada fames ac ante ipsum primis in faucibus. Sed elementum risus tellus, nec pharetra tortor feugiat id. Fusce pretium tincidunt efficitur. Praesent ut sodales neque. Duis imperdiet, risus vitae commodo fringilla, metus nibh egestas lectus, sed congue lorem enim nec ex. Duis nec porttitor diam. Donec aliquet metus eu ante cursus tempor. Nunc malesuada mi venenatis, tincidunt est suscipit, tempor nunc. Sed scelerisque erat id magna dictum, ut posuere mauris sodales. Suspendisse interdum posuere tellus, ut commodo metus sagittis ut.',0);";
+      $query="INSERT INTO todo (done_todo,objectif_todo) VALUES
+      (0,'Integer at odio porta, finibus mi vel, bibendum orci. Duis elit est, ornare in ornare eu, dictum et nulla. Pellentesque libero enim, consectetur sed lectus vitae, tristique interdum nisi. Suspendisse eget tempor erat, et cursus risus. Nam purus velit, facilisis a placerat tempus, consequat vitae mi. Nulla rhoncus, quam non imperdiet malesuada, erat enim sollicitudin nisi, in interdum orci mauris sit amet leo. Donec sagittis ligula ipsum, et pharetra augue porta ut. Donec vel posuere quam. Phasellus facilisis ligula a sem pretium, eu consequat nibh mattis. Fusce id turpis eget elit maximus laoreet eget nec turpis. Duis id sem lobortis, ornare turpis in, tempor augue. Donec at magna vel leo varius ullamcorper non ut elit.'),
+      (1,'Nulla facilisi. Maecenas tristique dolor in enim porttitor laoreet. Duis venenatis tincidunt purus, sed suscipit lorem rutrum ac. Proin mauris massa, blandit sit amet porttitor fermentum, pharetra sit amet diam. Aliquam id pulvinar orci, id condimentum lorem. Interdum et malesuada fames ac ante ipsum primis in faucibus. Sed elementum risus tellus, nec pharetra tortor feugiat id. Fusce pretium tincidunt efficitur. Praesent ut sodales neque. Duis imperdiet, risus vitae commodo fringilla, metus nibh egestas lectus, sed congue lorem enim nec ex. Duis nec porttitor diam. Donec aliquet metus eu ante cursus tempor. Nunc malesuada mi venenatis, tincidunt est suscipit, tempor nunc. Sed scelerisque erat id magna dictum, ut posuere mauris sodales. Suspendisse interdum posuere tellus, ut commodo metus sagittis ut.'),
+      (0,'Nulla facilisi. Maecenas tristique dolor in enim porttitor laoreet. Duis venenatis tincidunt purus, sed suscipit lorem rutrum ac. Proin mauris massa, blandit sit amet porttitor fermentum, pharetra sit amet diam. Aliquam id pulvinar orci, id condimentum lorem. Interdum et malesuada fames ac ante ipsum primis in faucibus. Sed elementum risus tellus, nec pharetra tortor feugiat id. Fusce pretium tincidunt efficitur. Praesent ut sodales neque. Duis imperdiet, risus vitae commodo fringilla, metus nibh egestas lectus, sed congue lorem enim nec ex. Duis nec porttitor diam. Donec aliquet metus eu ante cursus tempor. Nunc malesuada mi venenatis, tincidunt est suscipit, tempor nunc. Sed scelerisque erat id magna dictum, ut posuere mauris sodales. Suspendisse interdum posuere tellus, ut commodo metus sagittis ut.')";
       $this->remove_all_todo();
       return connect_sql($query);
     }
@@ -24,6 +27,9 @@
       return connect_sql($query);
     }
     function remove_todo(string $id){
+      if(!check_id($id)){
+        return;
+      }
       $query="delete from todo where id_todo=$id";
       return connect_sql($query);
     }
@@ -41,8 +47,14 @@
       return connect_sql($query);
     }
     function modify_todo(string $id, string $newvalue){
+      if(!check_id($id)){
+        return;
+      }
       $query="update todo set objectif_todo='$newvalue' where id_todo=$id";
       return connect_sql($query);
     }
+  }
+  function check_id(string $id):bool{
+    return ctype_digit($id);
   }
 ?>
